requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.12.0
numpy==1.23.5
protobuf==3.20.3
pillow
opencv-python-headless